requirements.txt
numpy>=1.24
pandas>=2.0
requests>=2.31
pyyaml>=6.0
streamlit>=1.33
google-generativeai>=0.8.0
